fix(skills): decouple parseFrontmatter from cli.js for CI validator
badi-skills CI was failing because the validate workflow curls schema.js
and icerik-helpers.js but icerik-helpers.js imports chalk from cli.js,
which the workflow does not fetch (and pulling cli.js would also need
package.json + chalk + figlet).

Move parseFrontmatter to a zero-dependency module (lib/frontmatter.js)
and have schema.js import directly from there. icerik-helpers.js
re-exports parseFrontmatter for backwards compatibility.

Verified locally with the same CI flow (curl-equivalent copy + sed
import path rewrite + validate.mjs against a sample SKILL.md): module
loads cleanly and validation passes.
