+++ lib/frontmatter.js
@@ -1,0 +1,28 @@
+// Minimal markdown frontmatter parser.
+//
+// Bagimsiz tutuldu (chalk/cli yok) — boylece harici tooling (orn.
+// badi-skills CI validate workflow) sadece bu dosyayi cekerek skill
+// schema'sini dogrulayabilir.
+
+export function parseFrontmatter(content) {
+	const lines = content.split("\n");
+	if (lines[0] !== "---") return { meta: {}, body: content };
+	const endIdx = lines.indexOf("---", 1);
+	if (endIdx === -1) return { meta: {}, body: content };
+	const meta = {};
+	for (let i = 1; i < endIdx; i++) {
+		const colonIdx = lines[i].indexOf(": ");
+		if (colonIdx > 0) {
+			meta[lines[i].substring(0, colonIdx).trim()] = lines[i]
+				.substring(colonIdx + 2)
+				.trim();
+		}
+	}
+	return {
+		meta,
+		body: lines
+			.slice(endIdx + 1)
+			.join("\n")
+			.trim(),
+	};
+}
