+++ lib/icerik-helpers.js
@@ -8,6 +8,9 @@ import {
 import { homedir } from "node:os";
 import { join } from "node:path";
 import { chalk } from "./cli.js";
+import { parseFrontmatter } from "./frontmatter.js";
+
+export { parseFrontmatter };
 
 export function loadPreferences() {
 	const prefFile = join(homedir(), ".config", "badi", "preferences.json");
@@ -80,29 +83,6 @@ export function calculateSimilarity(a, b) {
 	const union = new Set([...wordsA, ...wordsB]).size;
 	const jaccard = union > 0 ? intersection / union : 0;
 	return levSim * 0.4 + jaccard * 0.6;
-}
-
-export function parseFrontmatter(content) {
-	const lines = content.split("\n");
-	if (lines[0] !== "---") return { meta: {}, body: content };
-	const endIdx = lines.indexOf("---", 1);
-	if (endIdx === -1) return { meta: {}, body: content };
-	const meta = {};
-	for (let i = 1; i < endIdx; i++) {
-		const colonIdx = lines[i].indexOf(": ");
-		if (colonIdx > 0) {
-			meta[lines[i].substring(0, colonIdx).trim()] = lines[i]
-				.substring(colonIdx + 2)
-				.trim();
-		}
-	}
-	return {
-		meta,
-		body: lines
-			.slice(endIdx + 1)
-			.join("\n")
-			.trim(),
-	};
 }
 
 export function slugify(text) {
