+++ lib/skills/schema.js
@@ -11,7 +11,7 @@
 // cikti dili Turkce ama schema mesajlari (validate ciktisi) standart
 // sekilde okumak icin English.
 
-import { parseFrontmatter } from "../icerik-helpers.js";
+import { parseFrontmatter } from "../frontmatter.js";
 
 export const KNOWN_TOOLS = [
 	"Read",
